Refactor to core: initial commit. - remove tests - refactor /models to core/models - refactor /services to core/services
