+++ src/app/applications/app-list/app-list.component.ts
@@ -3,9 +3,9 @@ import { Subject } from 'rxjs';
 import { takeUntil } from 'rxjs/operators';
 import * as _ from 'lodash';
 
-import { Application } from 'app/models/application';
-import { ApplicationService } from 'app/services/application.service';
-import { CommentPeriodService } from 'app/services/commentperiod.service';
+import { Application } from 'core/models/application';
+import { ApplicationService } from 'core/services/application.service';
+import { CommentPeriodService } from 'core/services/commentperiod.service'
 import { ConstantUtils, CodeType } from 'app/utils/constants/constantUtils';
 
 const LIST_PAGE_SIZE = 10;
