+++ src/app/applications/app-map/marker-popup/marker-popup.component.ts
@@ -2,10 +2,10 @@ import { Component, OnInit, OnDestroy } from '@angular/core';
 import { Subject } from 'rxjs';
 // import { takeUntil } from 'rxjs/operators';
 
-import { Application } from 'app/models/application';
-import { ApplicationService } from 'app/services/application.service';
-import { CommentPeriodService } from 'app/services/commentperiod.service';
-import { UrlService } from 'app/services/url.service';
+import { Application } from 'core/models/application';
+import { ApplicationService } from 'core/services/application.service';
+import { CommentPeriodService } from 'core/services/commentperiod.service'
+import { UrlService } from 'core/services/url.service';
 import { Panel } from 'app/applications/utils/panel.enum';
 import { singleApplicationStub } from 'app/applications/stubs/application-stub';
