+++ src/core/models/document.ts
