+++ src/core/models/search.ts
