+++ src/core/services/decision.service.ts
@@ -2,8 +2,8 @@ import { Injectable } from '@angular/core';
 import { Observable, of } from 'rxjs';
 import { map, mergeMap, catchError } from 'rxjs/operators';
 
-import { ApiService } from './api';
-import { Decision } from 'app/models/decision';
+import { ApiService } from 'core/services/api';
+import { Decision } from 'core/models/Decision';
 import { DocumentService } from './document.service';
 
 @Injectable()
